sqrt overflow handled 2

+++ BinarySearch/sqrt.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class sqrt {
     public static void main(String args[]){
         Scanner sc = new Scanner(System.in);
-        int x = sc.nextInt();
-        int s = 1, e = x, m;
-        int ans = 1;
+        long x = sc.nextInt();
+        long s = 1, e = x, m;
+        long ans = 1;
         while(s <= e){
             m = (s + e)/2;
             if(m <= (x/m)){
